// src/dto/plant.dto.ts
export class PlantDTO {
  facIdentifier: number;

  orisCode: number;

  facilityName: string;

  description: string;

  state: string;

  countyCode: string;

  sicCode: number;

  epaRegion: number;

  nercRegion: string;

  airsid: string;

  findsid: string;

  stateid: string;

  latitude: number;

  longitude: number;

  frsIdentifier: string;

  payeeIdentifier: number;

  permitEXPDate: string;

  latlonSource: string;

  tribalLandCode: string;

  firstEcmpsRPTPeriodIdentifier: number;

  userid: string;

  addDate: string;

  updateDate: string;
}